Added screen navigation, minor fixes.
Screens can now easily be navigated back and forth, via a parsed
"previous" Display value, as a part of a Display's constructor.
Increased the size of the Display from 150x40 to 151x41, to ensure
there's a Pxl that's directly in the middle of the display.
Scrapped Menus, as Prefabs have replaced them.
Forced resource rendering to be converted to UTF8.
Added RenderMenuExit prefab.