Display.Write center, Write ignores '~'
Display.Write now works like Display.DrawResource; its posX and posY parameters can be null, in which case it will go for the center in the respective dimensions.
Display.WriteCentered removed, as Display.Write can now do the same.
Display.Write now ignores the character '~' similarly to Display.DrawResource; if it is seen in a string, it will not replace whichever Pxl was rendered on that spot of the grid.